some tests were added, CommonService.deleteById() now returns boolean

+++ src/main/java/com/toolsapp/service/AbstractCommonService.java
@@ -41,7 +41,13 @@ public abstract class AbstractCommonService
         return entity.get();
     }
 
-    public void deleteById(long id) {
-        repository.deleteById(id);
+    public boolean deleteById(long id) {
+        if (findById(id).isPresent()) {
+            repository.deleteById(id);
+            return true;
+        }
+        else {
+            return false;
+        }
     }
 }
